Fix behavior when user data trace is not saved
Previously, if the user data trace was not being saved, the output was not returned at all in the objective function. This caused the BayesOpt to run the objective function once, just to figure out how many outputs it gives. Now the objective function provides the user data output in either case, but it is empty if we don't want to store it.