Modify fit_models to support sequential execution
Modified fit_models such that it perform a for-loop when parallel execution is set to `false`, and to perform a parfor-loop when parallel execution is set to `true`. Previously, the serial execution was done using the parfor with number of workers set to 0, but this still results in non-deterministic execution (and it may depend on the parallel computing toolbox (?) ).